outputPackageInfo
