+++ libs/@local/repo-chores/scripts/shared/package-infos.ts
@@ -1,7 +1,9 @@
 import path from "node:path";
 
+import * as envalid from "envalid";
 import fs from "fs-extra";
 
+import { UserFriendlyError } from "./errors";
 import { monorepoRoot } from "./monorepo-root";
 
 export interface PackageInfo {
@@ -46,4 +48,38 @@ export const listPublishablePackages = async (): Promise<PackageInfo[]> => {
   }
 
   return result;
+};
+
+export const derivePackageInfoFromEnv = async (): Promise<PackageInfo> => {
+  const env = envalid.cleanEnv(process.env, {
+    PACKAGE_DIR: envalid.str({
+      desc: "location of package",
+    }),
+  });
+  const packageDirPath = path.resolve(env.PACKAGE_DIR);
+
+  if (packageDirPath !== env.PACKAGE_DIR) {
+    throw new UserFriendlyError(
+      `PACKAGE_DIR must be an absolute path, got ${packageDirPath}`,
+    );
+  }
+
+  const publishablePackageInfos = await listPublishablePackages();
+  const packageInfo = publishablePackageInfos.find(
+    (currentPackageInfo) => currentPackageInfo.path === packageDirPath,
+  );
+
+  if (!packageInfo) {
+    throw new UserFriendlyError(
+      `PACKAGE_DIR (${packageDirPath}) does not point to a publishable package`,
+    );
+  }
+  return packageInfo;
+};
+
+export const outputPackageInfo = (packageInfo: PackageInfo): void => {
+  console.log("");
+  console.log(`Package name: ${packageInfo.name}`);
+  console.log(`Package path: ${packageInfo.path}`);
+  console.log("");
 };
